feat: refactor bot.js
