+++ config.js
@@ -1,0 +1,105 @@
+require('dotenv').config({ path: __dirname + '/.env' });
+const { createClient } = require('@supabase/supabase-js');
+const { Web3 } = require('web3');
+const fs = require('fs');
+
+// Supabase setup
+const supabase = createClient(
+  process.env.SUPABASE_URL,
+  process.env.SUPABASE_ANON_KEY
+);
+
+// Initialize web3
+const web3 = new Web3(process.env.BASE_RPC);
+
+// Load Escrow contract ABI
+const ESCROW_ABI = JSON.parse(fs.readFileSync(__dirname + '/abi.js', 'utf8'));
+
+// Initialize escrow contract
+const escrowContract = new web3.eth.Contract(ESCROW_ABI, '0xca38607d85e8f6294dc10728669605e6664c2d70');
+
+// Exchange rate API configuration
+const EXCHANGE_API_URL = `https://v6.exchangerate-api.com/v6/${process.env.EXCHANGE_API_KEY}/latest/USD`;
+
+// Global storage
+const depositAmounts = new Map(); // Store deposit amounts temporarily
+const intentDetails = new Map();
+
+// Currency and platform mappings
+const verifierMapping = {
+  '0x8b5d4e8768d7b05e8f2a8e4b7b8b2d7e8c8d7a6b5c4e3f2a1b9c8d7e6f5a4': 'revolut',
+  '0x7a6b5c4e3f2a1b9c8d7e6f5a4b3c2d1e9f8a7b6c5d4e3f2a1b9c8d7e6f5a4': 'wise',
+  '0x6b5c4e3f2a1b9c8d7e6f5a4b3c2d1e9f8a7b6c5d4e3f2a1b9c8d7e6f5a4b3': 'paypal',
+  '0x5c4e3f2a1b9c8d7e6f5a4b3c2d1e9f8a7b6c5d4e3f2a1b9c8d7e6f5a4b3c2': 'venmo',
+  '0x4e3f2a1b9c8d7e6f5a4b3c2d1e9f8a7b6c5d4e3f2a1b9c8d7e6f5a4b3c2d1': 'zelle',
+  '0x3f2a1b9c8d7e6f5a4b3c2d1e9f8a7b6c5d4e3f2a1b9c8d7e6f5a4b3c2d1e9': 'ach',
+};
+
+const getPlatformName = (verifierAddress) => {
+  return verifierMapping[verifierAddress] || 'Unknown Platform';
+};
+
+const currencyHashToCode = {
+  '0x4555520000000000000000000000000000000000000000000000000000000000': 'EUR',
+  '0x4742500000000000000000000000000000000000000000000000000000000000': 'GBP',
+  '0x4a50590000000000000000000000000000000000000000000000000000000000': 'JPY',
+  '0x4155440000000000000000000000000000000000000000000000000000000000': 'AUD',
+  '0x4341440000000000000000000000000000000000000000000000000000000000': 'CAD',
+  '0x4348460000000000000000000000000000000000000000000000000000000000': 'CHF',
+  '0x434e590000000000000000000000000000000000000000000000000000000000': 'CNY',
+  '0x494e520000000000000000000000000000000000000000000000000000000000': 'INR',
+  '0x4b52570000000000000000000000000000000000000000000000000000000000': 'KRW',
+  '0x42524c0000000000000000000000000000000000000000000000000000000000': 'BRL',
+  '0x4d584e0000000000000000000000000000000000000000000000000000000000': 'MXN',
+  '0x5347440000000000000000000000000000000000000000000000000000000000': 'SGD',
+  '0x4e5a440000000000000000000000000000000000000000000000000000000000': 'NZD',
+  '0x5a41520000000000000000000000000000000000000000000000000000000000': 'ZAR',
+  '0x4855460000000000000000000000000000000000000000000000000000000000': 'HUF',
+  '0x504c4e0000000000000000000000000000000000000000000000000000000000': 'PLN',
+  '0x435a4b0000000000000000000000000000000000000000000000000000000000': 'CZK',
+  '0x444b4b0000000000000000000000000000000000000000000000000000000000': 'DKK',
+  '0x4e4f4b0000000000000000000000000000000000000000000000000000000000': 'NOK',
+  '0x53454b0000000000000000000000000000000000000000000000000000000000': 'SEK',
+  '0x494c530000000000000000000000000000000000000000000000000000000000': 'ILS',
+  '0x5448440000000000000000000000000000000000000000000000000000000000': 'THB',
+  '0x4944520000000000000000000000000000000000000000000000000000000000': 'IDR',
+  '0x4d594d0000000000000000000000000000000000000000000000000000000000': 'MYR',
+  '0x5048500000000000000000000000000000000000000000000000000000000000': 'PHP',
+  '0x564e440000000000000000000000000000000000000000000000000000000000': 'VND',
+  '0x4547590000000000000000000000000000000000000000000000000000000000': 'EGY',
+  '0x4b45530000000000000000000000000000000000000000000000000000000000': 'KES',
+  '0x4e474e0000000000000000000000000000000000000000000000000000000000': 'NGN',
+  '0x4748530000000000000000000000000000000000000000000000000000000000': 'GHS',
+  '0x54415a0000000000000000000000000000000000000000000000000000000000': 'TZS',
+  '0x5547580000000000000000000000000000000000000000000000000000000000': 'UGX',
+};
+
+const formatConversionRate = (conversionRate, fiatCode) => {
+  if (!conversionRate || !fiatCode) return 'N/A';
+
+  const rate = parseFloat(conversionRate);
+  if (isNaN(rate)) return 'N/A';
+
+  // Format based on currency
+  const decimalCurrencies = ['JPY', 'KRW', 'VND', 'IDR'];
+  if (decimalCurrencies.includes(fiatCode)) {
+    return Math.round(rate).toLocaleString();
+  } else {
+    return rate.toFixed(2);
+  }
+};
+
+module.exports = {
+  supabase,
+  web3,
+  escrowContract,
+  EXCHANGE_API_URL,
+  depositAmounts,
+  intentDetails,
+  verifierMapping,
+  getPlatformName,
+  currencyHashToCode,
+  formatConversionRate,
+  CONTRACT_ADDRESS: process.env.CONTRACT_ADDRESS,
+  WEBHOOK_URL: process.env.WEBHOOK_URL
+};
